Refactored various interfaces.

+++ sources/TCPSocket.cpp
@@ -30,12 +30,12 @@ SocketProtocol TCPSocket::Protocol() const
 
 int TCPSocket::Send(const std::string& s)
 {
-    return Send(s.c_str(), static_cast<int>(s.size()));
+    return Send(s.c_str(), s.size());
 }
 
 int TCPSocket::Recv(std::string& s)
 {
-    auto result = Recv(&s[0], static_cast<int>(s.size()));
+    auto result = Recv(&s[0], s.size());
     if (result >= 0)
         s.resize(static_cast<std::size_t>(result));
     return result;
