+++ sources/UDPSocket.cpp
@@ -30,12 +30,12 @@ SocketProtocol UDPSocket::Protocol() const
 
 int UDPSocket::Send(const std::string& s, const IPAddress& address)
 {
-    return Send(s.c_str(), static_cast<int>(s.size()), address);
+    return Send(s.c_str(), s.size(), address);
 }
 
 int UDPSocket::Recv(std::string& s, IPAddress& address)
 {
-    auto result = Recv(&s[0], static_cast<int>(s.capacity()), address);
+    auto result = Recv(&s[0], s.capacity(), address);
     if (result >= 0)
         s.resize(static_cast<std::size_t>(result));
     return result;
